Update __init__.py
added import of requests so that this actually run.
perhaps there are other packaging methods you are using where this isn't an issue but I'm not aware of it.

Also, it would be very helpful to move this into https://pypi.org/
to make this more generally available.   I'm curious as to why this has not happened?
If I can help in anyway to make this happen I'm willing to do so.